requirements.txt
annotated-types==0.6.0
anyio==4.3.0
atproto==0.0.46
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cohere==5.3.4
cryptography==42.0.7
distro==1.9.0
dnspython==2.6.1
exceptiongroup==1.2.1
fastavro==1.9.4
filelock==3.14.0
fsspec==2024.3.1
h11==0.14.0
httpcore==1.0.5
httpx==0.25.2
httpx-sse==0.4.0
huggingface-hub==0.23.0
idna==3.7
libipld==1.2.3
openai==1.26.0
packaging==24.0
pip-review==1.3.0
praw==7.7.1
prawcore==2.4.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.1
tokenizers==0.19.1
tqdm==4.66.4
types-requests==2.31.0.20240406
typing_extensions==4.11.0
update-checker==0.18.0
urllib3==2.2.1
websocket-client==1.8.0
websockets==12.0
